requirements.txt
# Core
python-dotenv>=0.19.0
pyyaml>=6.0
pandas>=1.3.0
numpy>=1.21.0

# Web Scraping
requests>=2.26.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
webdriver-manager>=3.5.2

# Natural Language Processing
nltk>=3.6.3
spacy>=3.2.0
textblob>=0.17.1

# Database
cx_Oracle>=8.3.0
SQLAlchemy>=1.4.29

# Data Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Jupyter
jupyter>=1.0.0
ipykernel>=6.4.1

# Code Quality
black>=21.12b0
flake8>=4.0.1
isort>=5.10.1

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
